# GATEWAY_OPENAPI.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PaddyGuard AI \u2014 API Gateway",
    "description": "Central gateway for all PaddyGuard AI microservices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/chat/message": {
      "post": {
        "tags": [
          "Treatment Chatbot \u2014 C4"
        ],
        "summary": "Send Message",
        "description": "Forward chat message to C4 RAG chatbot.",
        "operationId": "send_message_api_v1_chat_message_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Payload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat/session/{session_id}": {
      "delete": {
        "tags": [
          "Treatment Chatbot \u2014 C4"
        ],
        "summary": "Clear Session",
        "description": "Forward to C4's DELETE /chat/session/:sessionId \u2014 used when the user starts a new chat.",
        "operationId": "clear_session_api_v1_chat_session__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [
    {
      "url": "https://gateway-production-3c64.up.railway.app"
    }
  ]
}